связи таблиц

+++ src/main/java/ru/skypro/homework/store/entities/CommentEntity.java
@@ -18,7 +18,7 @@ public class CommentEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Long id;
+    int id;
 
     @Column(name = "count", nullable = false)
     int count;
